+++ ChangeLog.md
@@ -20,6 +20,15 @@
 
 # Change Log
 
+## Upcoming Release
+
+* Invoices 
+  * Renamed the reconciliation line item objects
+* Users
+  * No longer throw an error when searching for a user with UPN that does not exists. 
+* Utilization 
+  * Modified the default end date value for the Get-PartnerCustomerSubscriptionUtilization command to use UTC time.
+
 ## 1.5.1903.6
 
 * Products